+++ spec/components/related_navigation_spec.rb
@@ -155,10 +155,27 @@ describe "Related navigation", type: :view do
     render_component(content_item:)
 
     assert_select ".gem-c-related-navigation__section-link[href=\"/world/wales/news\"]", text: "Wales"
-    assert_select ".gem-c-related-navigation__link.toggle-wrap", text: "Show 2 more"
+    assert_select ".gem-c-related-navigation__link.toggle-wrap", text: "Show 2 more world locations"
     assert_select ".gem-c-related-navigation__link.toggle-wrap[data-module='ga4-event-tracker']"
     assert_select "#toggle_world_locations .gem-c-related-navigation__section-link[href=\"/world/mauritius/news\"]", text: "Mauritius"
     assert_select "#toggle_world_locations .gem-c-related-navigation__section-link[href=\"/world/brazil/news\"]", text: "Brazil"
+  end
+
+  it "adds a show more toggle link to long external link sections" do
+    content_item = {}
+    links = []
+
+    7.times do
+      links << { "url" => "https://wikipedia.org", "title" => "Wikipedia" }
+    end
+
+    content_item["details"] = {
+      "external_related_links" => links,
+    }
+
+    render_component(content_item:)
+
+    assert_select ".gem-c-related-navigation__link.toggle-wrap", text: "Show 2 more external links"
   end
 
   it "does not use a Show More for only one link above the max per section" do
@@ -250,7 +267,7 @@ describe "Related navigation", type: :view do
 
     assert_select ".gem-c-related-navigation__section-link[href=\"/world/wales/news\"]", text: "Wales"
     assert_select ".gem-c-related-navigation__link.toggle-wrap[data-module='ga4-event-tracker']"
-    assert_select ".gem-c-related-navigation__toggle[data-ga4-event='{\"event_name\":\"select_content\",\"type\":\"related content\"}']", text: "Show 2 more"
+    assert_select ".gem-c-related-navigation__toggle[data-ga4-event='{\"event_name\":\"select_content\",\"type\":\"related content\"}']", text: "Show 2 more world locations"
   end
 
   it "allows GA4 to be disabled" do